identification/requirements.txt
joblib
numpy
pandas
praat-parselmouth
pyaudio
python_speech_features
scipy
sklearn
scikit-learn
tensorflow
IPython
keras